requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.6
msal==1.26.0
numpy==1.26.3
Office365-REST-Python-Client==2.5.4
openpyxl==3.1.2
pandas==2.1.4
pandasql==0.7.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pythonping==1.1.4
pytz==2023.3.post1
pywin32==306
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
WMI==1.5.1
xlrd==2.0.1
XlsxWriter==3.1.9
